fix: strike와 ball count가 이전 출력값에서 계속 더해지는 오류 수정
BallStrike 클래스를 따로 두고 BaseballCount에서
countStrike메서드와 countBall메서드에 지역변수를 생성하여 해결
